requirements.txt
langchain
langchain-openai
fastapi
elasticsearch
python-dotenv
uvicorn
pydantic
requests
python-jose
python-multipart
aiohttp
pywin32; platform_system == "Windows"
pytest
pytest-asyncio
nest-asyncio
httpx